refactor: config, prisma, middleware → common 으로 이동
