+++ src/app.module.ts
@@ -3,11 +3,11 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { AuthModule } from './auth/auth.module';
-import authConfig from './config/auth.config';
-import emailConfig from './config/email.config';
-import { validationSchema } from './config/validation.schema';
+import authConfig from './common/config/auth.config';
+import emailConfig from './common/config/email.config';
+import { validationSchema } from './common/config/validation.schema';
 import { HealthCheckModule } from './health-check/health-check.module';
-import { PrismaModule } from './prisma/prisma.module';
+import { PrismaModule } from './common/prisma/prisma.module';
 import { UsersModule } from './users/users.module';
 import { EventEmitterModule } from '@nestjs/event-emitter';
