+++ src/auth/auth.service.ts
@@ -11,8 +11,8 @@ import * as bcrypt from 'bcrypt';
 import * as jwt from 'jsonwebtoken';
 import { UserCreatedEvent } from 'src/auth/event/user-created.event';
 import { EventType } from 'src/common/event.type';
-import authConfig from 'src/config/auth.config';
-import { PrismaService } from 'src/prisma/prisma.service';
+import authConfig from 'src/common/config/auth.config';
+import { PrismaService } from 'src/common/prisma/prisma.service';
 import * as uuid from 'uuid';
 import { LoginUserDto } from './dto/login-user.dto';
 import { SignupUserDto } from './dto/signup-user.dto';
