+++ src/email/email.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { ConfigType } from '@nestjs/config';
 import Mail from 'nodemailer/lib/mailer';
-import emailConfig from 'src/config/email.config';
+import emailConfig from 'src/common/config/email.config';
 import * as nodemailer from 'nodemailer';
 
 interface EmailOptions {
